requirements.txt
Flask==0.11
Flask-Script==2.0.5
pymongo==3.4.0
mongoengine==0.11.0
flask-mongoengine==0.8.2
py-bcrypt==0.4
jsonschema==2.5.1
pyjwt==1.4.2